java/src/test/java/com/yourtion/leetcode/daily/m06/d06/SolutionTest.java
package com.yourtion.leetcode.daily.m06.d06;

import com.yourtion.leetcode.TestUtils;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.params.provider.Arguments.arguments;

@DisplayName("每日一题 - 20200606")
class SolutionTest {

    static Stream<Arguments> testDataProvider() {
        return Stream.of(
                arguments("[100, 4, 200, 1, 3, 2]", 4)
        );
    }

    @ParameterizedTest()
    @MethodSource("testDataProvider")
    void longestConsecutive(String source, int res) {
        System.out.printf("runTest: %s , res: %d", source, res);
        int[] ss = TestUtils.stringToIntegerArray(source);
        Assertions.assertEquals(res, new Solution().longestConsecutive(ss));
    }
}